requirements.txt
agentql
playwright
python-dotenv
requests
pandas
matplotlib
seaborn 